gityping: Remove inaccurate comment
It's hard to know what I was thinking, but I feel like this was from a
time when we *were* actually returning classes here and things might be
weird. Now, we return strings for all class parameters. I've checked
functions with return types with overrides and parameters that are
overrides, and so far everything looks as I'd expect, so... yeah.